TTextDevice: remove buffering and fix 'do_sputn' instead
When buffering was added in 0556fa4fd6c9b5a134407769fb157043ad3f4327, we were trying to fix the performance issue caused by 'do_sputn' being only passed one character at a time but redrawing the whole TTerminal each time.

However, such buffering was broken, so it had to be disabled for Borland C++ in 597c801eeb19a76cc4e6df58b63339520bedfd55.

It turns out that STL's streambuf and Borland's streambuf work differently. For example, 'streambuf::sputn' just invokes 'xsputn' in the STL while it may write to the streambuf buffer in Borland's RTL. As a result, properly implementing buffering via the streambuf interface would require altering the TTextDevice's public interface and increasing code complexity.

What can be easily done is to fix the usage of 'do_sputn' so that it gets several characters at a time, as it does in Borland C++. This is enough to solve the original issue.