Avoid complex memory allocations in TTerminal::draw
A dynamically-allocated array was added in ee821b69c5dd81c565fe1add1ac6f0a2f8a96a01 in order to support long lines and to avoid truncating multibyte characters. However, the resulting code was messy. We can simplify it by using a fixed-sized array in which we make sure that no characters get truncated. We can also use a proper TDrawBuffer instead of alloca.

In 16-bit mode this fixes a big memory safety issue of the original code, which would cause a buffer overrun when processing a line longer than 256 characters.